Replace notifications with workout page and update account details
