+++ app/src/main/java/com/rumiznellasery/yogahelper/ui/dashboard/DashboardFragment.java
@@ -1,6 +1,5 @@
 package com.rumiznellasery.yogahelper.ui.dashboard;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -20,16 +19,6 @@ public class DashboardFragment extends Fragment {
         binding = FragmentDashboardBinding.inflate(inflater, container, false);
 
         View root = binding.getRoot();
-
-        View.OnClickListener listener = v -> {
-            Intent intent = new Intent(requireContext(), com.rumiznellasery.yogahelper.temp.TempActivity.class);
-            startActivity(intent);
-        };
-
-        binding.buttonPlaceholder1.setOnClickListener(listener);
-        binding.buttonPlaceholder2.setOnClickListener(listener);
-        binding.buttonPlaceholder3.setOnClickListener(listener);
-
         return root;
     }
