+++ app/src/main/java/com/rumiznellasery/yogahelper/ui/home/HomeFragment.java
@@ -4,7 +4,8 @@ import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
 
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
@@ -23,6 +24,18 @@ public class HomeFragment extends Fragment {
 
         binding = FragmentHomeBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
+
+        FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
+        if (user != null) {
+            StringBuilder sb = new StringBuilder();
+            if (user.getDisplayName() != null) sb.append("Name: ").append(user.getDisplayName()).append('\n');
+            if (user.getEmail() != null) sb.append("Email: ").append(user.getEmail()).append('\n');
+            if (user.getPhoneNumber() != null) sb.append("Phone: ").append(user.getPhoneNumber()).append('\n');
+            sb.append("UID: ").append(user.getUid());
+            binding.textAccountDetails.setText(sb.toString());
+        } else {
+            binding.textAccountDetails.setText("No account data");
+        }
 
         return root;
     }
