+++ app/src/main/java/com/rumiznellasery/yogahelper/ui/workout/WorkoutViewModel.java
@@ -1,16 +1,16 @@
-package com.rumiznellasery.yogahelper.ui.notifications;
+package com.rumiznellasery.yogahelper.ui.workout;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class NotificationsViewModel extends ViewModel {
+public class WorkoutViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public NotificationsViewModel() {
+    public WorkoutViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("You Have No Notifications :)");
+        mText.setValue("Workout Page");
     }
 
     public LiveData<String> getText() {
