+++ app/src/main/res/layout/fragment_dashboard.xml
@@ -1,88 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <LinearLayout
-        android:layout_width="match_parent"
+    <TextView
+        android:id="@+id/text_dashboard"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:padding="16dp">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="16dp"
-            android:orientation="vertical"
-            android:background="@drawable/rounded_box"
-            android:padding="16dp">
-
-            <TextView
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="Placeholder 1"
-                android:textColor="@color/black"
-                android:layout_marginBottom="8dp" />
-
-            <Button
-                android:id="@+id/button_placeholder1"
-                android:layout_width="120dp"
-                android:layout_height="48dp"
-                android:text="Start"
-                android:textColor="@color/white"
-                android:background="@drawable/circular_button"
-                android:layout_gravity="center_horizontal" />
-        </LinearLayout>
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="16dp"
-            android:orientation="vertical"
-            android:background="@drawable/rounded_box"
-            android:padding="16dp">
-
-            <TextView
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="Placeholder 2"
-                android:textColor="@color/black"
-                android:layout_marginBottom="8dp" />
-
-            <Button
-                android:id="@+id/button_placeholder2"
-                android:layout_width="120dp"
-                android:layout_height="48dp"
-                android:text="Start"
-                android:textColor="@color/white"
-                android:background="@drawable/circular_button"
-                android:layout_gravity="center_horizontal" />
-        </LinearLayout>
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            android:background="@drawable/rounded_box"
-            android:padding="16dp">
-
-            <TextView
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="Placeholder 3"
-                android:textColor="@color/black"
-                android:layout_marginBottom="8dp" />
-
-            <Button
-                android:id="@+id/button_placeholder3"
-                android:layout_width="120dp"
-                android:layout_height="48dp"
-                android:text="Start"
-                android:textColor="@color/white"
-                android:background="@drawable/circular_button"
-                android:layout_gravity="center_horizontal" />
-        </LinearLayout>
-
-    </LinearLayout>
-</ScrollView>
+        android:text="Dashboard"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"/>
+</androidx.constraintlayout.widget.ConstraintLayout>
